backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.3.0
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.2
python-dotenv==1.0.0
